Created UserManager file
